feat: 컴포넌트 분리, util 함수 적용
