+++ src/app/(with-header)/mypage/components/stats-cards.tsx
@@ -21,7 +21,6 @@ export default function StatsCards({
   return (
     <div className="grid grid-cols-1 md:grid-cols-3 gap-6 mb-8">
       <motion.div
-        initial={{ y: 20, opacity: 0 }}
         animate={{ y: 0, opacity: 1 }}
         transition={{ delay: 0.1 }}
         onClick={() => onClickTab("submitted")}
